Simplify trait implementation
A trait is a specification of expected interface. If a trait defines
function foo(), a type implementing this trait must provide a function
foo().

The simplest traits require types that implement them to implement just
a single function. The name of this function is operator(), with a
special tag as its first parameter.

Definition of such traits may be emitted using a simple macro. Let's do
this instead of writing the code by hand.