Boxed versions of fundamental types in preparation for pointers
Pointers are checked and need cooperation from values they point to.
This is why pointers to fundamental values cannot be taken - because the
fundamental values are dumb, and cannot offer the level of cooperation
required to make pointers safe. At least until I write a sufficiently
smart static analyser, that is.