Rework storing stuff in and fetching stuff from registers
The last attempt was half-assed and broke down pretty quickly. This time
I will do better (at least that's what I am telling myself).

A register set is an array of cells, and each cell may be empty or hold
a value. The value may be of a primitive, unboxed type (int, float), or
of a boxed type (eg, a string or a struct).

Instructions which create new values or move values between registers
use the "slot" view of registers. For them it is not important WHAT
exactly the register contains, as they will just reset the slots
contents or shuffle them around.

Instructions which mutate existing values use the "cell" view of
registers. They are not concerned by WHERE exactly the cell is, as they
will only use its value.

Of course, there are instructions which view some of their operands
through the "slot" lens and some through the "cell view" lens. An
instruction may change its point of view operand by operand.