addi works with boxed values
It will mutate the value if the output contains a boxed value of
suitable type instead of overwriting it. Without this behaviour it would
not be possible to increment values to which a pointer was taken.

This also means that pointers are more like references... Guess I gotta
change the naming here.