Support dereferencing in more instructions and positions
Sometimes allowing dereferencing does not make much sense, but when it
does it should be allowed to avoid breaking consistency for no reason.

As a rule of thumb:

- all input operands should support dereferencing
- some output operands should support dereferencing, becoming mutating
  output operands

The "mutating output" operands are output operands whose behaviour is
altered when they are supplied as a dereference. Usually, an instruction
stores a completely new value in an output operand. However, when the
operand is a "mutating output" the instruction should mutate a value
instead.

This is required to maintain the illusion of working "on values" and not
"on registers" in presence of boxed types and references. Simply
overwriting a value to which a reference was taken would not update the
referenced value and the modification would not be visible through the
reference.