Implement comparison ops and traits
The Cmp trait is used for types with well-defined relationships and
ordering ie, types where ordering can always be defined. This is also
enforced by the type system - using C++'s std::strong_ordering as the
type of values produced by the Cmp trait.

The Eq trait is used for types whose precise ordering is not required
and whose equality is interesting instead. This is much simpler because
they only need to implement the == operator.

Cmp trait is used for cmp, lt, and gt instructions. It is also used as
fallback for the eq instruction. Eq trait is only used for eq
instruction.

Only primitive types are comparable. Complex types ie, structs and
vectors do not have either Cmp or Eq traits.