Add timing information to the test suite
This will let me monitor how fast the machine "feels" like. The timing
is skewed because tests are short and the time spent starting up and
shutting the VM down probably dwarfs actual execution time. And it
includes the time to assemble the source code which is also important.

In short, this is something more akin to "development cycle time" than a
pure "execution time". Still useful, I'd say.