Allow labels to refer to constant data defined outside of functions
This is mostly useful for strings. Imagine if you wanted to define a
buffer holding 4 KiB of memory. Defining it inside a function would
require several instructions or creating a looong string, like this:

    string $1, "        ...        "

Where the distance between the opening and closing " would be 4096 space
characters. It's obviously not ideal.

Labels and values solve this problem:

    .label: a_buffer
    .value: string " " * 4096

Voila!

We have a label "a_buffer" which refers to a value representing a string
of 4096 spaces. Simple.