Make main() of the assembler shorter
Move various stages of the assembly out to their own functions. This
also makes it clear which stage requires what, and what variables should
or could be grouped together in structs to make the whole setup easier
on the mind.

Source path and text are good candidates, as are function table and
offsets. Maybe in another commit.