Assembler writes ELF sections in the output
This should make locating the symbol table, text, and data segments
easier since they are all PT_LOAD and relying on ordering feels...
dirty. Now that sections are mapped, work can start on loading the ELF
in a civilised way and linking multiple files together.

The current code is REALLY BAD, though. It should be reworked to avoid
magic constants and make everything (or almost everything) automatic and
self-adjusting. This mostly concerns the number of segments and
sections.