Instead of hardcoding segments and sections, keep a list of them
This will require rethinking of how the offsets are calculated to keep
the code sane, but this is the direction the code SHOULD go. Hardcoding
may be easier in the beginning, but it becomes a royal pain in the ass
as the code outgrows initial assumptions.

And it WILL outgrow the initial assumptions, as it ALWAYS does...