Kinda-sorta improve reporting by readelf(1)
The mapping of segment 0 to section .viua.magic is now displayed
correctly, but readelf(1) decided to also map section .viua.magic in
segment 1. I don't really know why, and could not fix it. The output is
now as follows:

     Section to Segment mapping:
      Segment Sections...
       00     .viua.magic
       01     .viua.magic .interp

Kinda OK, kinda not. Anyway, when the VM itself will be checking the
file it will get the mapping correctly and be able to verify the file as
Viua binary. So the fact that readelf(1) produces weird output is not a
problem.