Keep segments and sections in a single vector
Avoid code duplication. Viua emits ELFs in which each segment is mapped
to its own section so it makes sense to keep them together - and it
makes the offset adjustment possible in one loop. There are only a few
exceptions to this rule, and even then it's just a section without a
mapped segment.

Keeping sections and segments in vectors also makes it easier to keep
the offset and size calculations automatic as almost nothing is (or can
be) hardcoded.