Fix the stage::emit_elf() to handle more cases correctly
Both object files and executables may contain an entry address. Viua
will use it to detect presence of the entry function, since relying on
the name of "main" will not be reliable (it can be overridden by setting
the [[entry_point]] tag on an arbitrarily named functtion).

The signature of stage::emit_elf() can be simplified, though. There is
no reason to pass the entry function as a lexeme instead of just an
offset - at this stage we don't really care about the name of the entry
function. Passing it as plain offset also allows us to drop the
parameter with function addresses.