Add basic loader frontend
The loader implementation will be reused in the VM, linker, and (in the
far future) debugger. The loader frontend serves as a way to analyse the
contents of Viua ELFs and can be used in tandem with readelf(1). It is a
tool specific to Viua so can emit output that is hopefully more relevant
to the VM.