Don't repack the loaded ELF file
The Env struct can just embed the Loaded_elf and feed off of its
contents. We just need to set the references correctly. There's still
the problem that the current prototype assumes only one module is needed
to execute the program, but this is something that can be solved in a
future patch.

Env's constructor assumes that the Loaded_elf it is passed represents an
ELF valid from Viua's point of view ie, that it contains the .rodata,
.text and .viua.fns sections (as the VM develops more sections will
probably be required). The VM's frontend verifies this and emits nice
error messages if it detects that the loaded ELF is not usable.