Allow expanding "short" integers using the canonical form
The canonical form is the full form to which li pseudoinstructions are
expanded for "wide" integers. Wide ie, integers that cannot fit into an
addi immediate. The canonical form is ALWAYS longer than a short form.
How to request the canonical form?

    li $1, [[full]] 42

Just use the [[full]] attribute before the immediate.

Why would someone request a longer sequence of instructions? The longer
form is useful for loading addresses. By always using the canonical form
we ensure that the linker will be able to patch the address with
arbitrary values (as long as fit into 64 bit wide register). Without
using the canonical form we could encounter situations where pre-patch
address can be loaded using fewer instructions then a post-patch
address. While not impossible to solve, it would require the linker to
shift all subsequent instructions and patch many more addresses. And we
don't wanna do this.