Add reassembly step to tests
Until now the tests were simple:

- assemble the source code
- run the executable ELF produced by the assembler
- check if ebreak output matches what was expected

The reassembly adds a few more step:

- disassemble the ELF
- run the initial steps on the disassembled source code

If such a test works OK, we can be reasonably sure that the whole
toolchain works correctly and you can freely assemble and reassemble the
source code and get equivallent ELFs.

Unfortunately, the disassembler does not yet fully work and running the
test suite with reassembly enabled wrecks the success rate. It dives
from 100% to just above 60%.