Demangle addiu instructions
Only ADDI is exposed to the user code, even if both ADDI and ADDIU are
implemented by the VM. The assembler distinguishes between signed and
unsigned integers and emits correct instructions based on type of the
value (this is consistent with runtime semantics, because ADD, MUL, etc.
also change behaviour based on the type of their operands).

Before this patch the disassembler was spitting out bullshit:

    addiu $1, void, X

And after this patch it speaks truth:

    addi $1, void, Xu

The `u' suffix means that the value is unsigned.