Prepare code for demangling aware of logical and physical positions
Instructions refer to logical instruction indexes on the user-visible
level, but switch to refering to physical indexes when compiled to
bytecode. The assembler handles translation of logical to physical
indexes, and the disassembler must reverse it.

Cooked instructions were up until now represented as a simple std::tuple
but this has outlived its usefulness. A time has come where a proper
struct with named fields is needed.