Release 0.11.1
The new, prototype distribution includes more or less functional tools
that allows somewhat normal workflow. There toolchain includes:

- assembler: for turning source code into executable ELFs
- disassembler: to make inspecting ELFs easier by converting them back
  into textual representation
- readelf: to easily get some high-level information about contents of
  ELFs produced by the assembler
- virtual machine kernel: for actually executing the bytecode contained
  in ELFs

There is no debugger (only a disassembler and an EBREAK instruction),
and no static analyser. The programmers are on their own and the VM went
back to living in the wild, wild west.

But I like the direction in which we the new prototype is going. I think
it will yield a better code then the VM I were developing for the
past... let me check my notes... 7? SEVEN FUCKING YEARS? Damn, how the
time flies. It was that long ago.

And still, no stable release. Nicely done, or rather NOT done. Good
going. Maybe I will have something usable when the VM hits a decade of
commit history. And that's a biiig maybe.