Add struct representing a process
A process has a stack, and the stack tracks IP (previously tracked as a
parameter of the instruction-running family of functions). The process
itself holds a pointer to core (to provide for I/O) and maintains
information about loaded modules (currently only one).

This split and reorganisation is needed to support more than one
concurrent processes in the new prototype.