Allow inspecting stacks of different processes in tests
The .ebreak files supplied in tests are now treated as scripts instead
of dumb dictionaries. Each line of an .ebreak script is evaluated and
verified against the "active" frame.

How is the active frame selected? Using the "ebreak" and "of" commands.
For example, to inspect the state of 2nd frame of the 3rd ebreak in
actor X you select it like this:

    ebreak 2 in process X
    of #1

You would then specify the expected register contents as usual ie, using
the following syntax:

    [1.l] is 000000000000002a 42

to say that local (because "l" is used) register 1 is expected to
contain a signed integer (the "is" type) with the value 42.

This makes what is being checked more explicit, and gives much more
control over the order of checks.