Fix the algorithm splitting and loading integers
There were two bugs, both related to integers which were so big that
they had a multiplier part, but so little that they had no high bits.
The "no high bits" part was the cause of both bugs.

First, the algorithm assumed that if a multiplier was present then it
could always add to the final destination register. This assumption was
false if no high bits were set since no LUI was emitted.

Second, the algorithm relies on the LUI to determine the final sign of
the loaded integer (since the left-hand side of an operation determines
type of the result).

This change fixes both bugs at once.