Basic ad-hoc evaluation capabilities in REPL
REPL is able to evaluate simple instructions ie, those which do not
refer to strings table. Pseudoinstructions are not supported yet, but
all moves, addis, muls, etc. should work OK.

Ad-hoc evaluation does not move ip of the executed process. The
instructions are executed from an ephemeral storage, but have lasting
effects on the frame in which they are executed.

This, in effect, allows tinkering with registers of suspended actors. A
more simple use for this functionality is to try out various
instructions and test their behaviour.