WIP: only throw Error values in assembler
Allowing Lexeme, Location, and Error to be thrown lead to multiple paths
for error handling that aimed at producing the same (or similar) error
reports.

Since the Error type is able to represent a broader range of errors than
bare Lexeme and Location it does not really make sense to use the other
types for error reporting... besides laziness.