Allow calling functions from REPL
The patch in execute(CALL) function is required because otherwise the
return addresses are all wrong. The REPL allocates memory for the code
to be evaluated dynamically, and deallocates after that code has been
executed. If the return address points to that now-deallocated memory
BAD things happen (ie, the IP becomes invalid).