Add AA, AD, and PTR instructions
They will be used to create pointers and allocate memory.

AA - Allocate Automatic

    aa $ptr, $size, <align>

Allocates memory with automatic duration (ie, on stack) and returns
pointer to the allocated memory area. The VM automatically finds a
suitable address to allocate the memory, including requesting a new page
if necessary.

The address parameter (2nd one) is used to specify the requested size.

The offset parameter (3rd one) is used to specify alignment. Alignment
of zero means the most stringent alignment is chosen.

AD - Allocate Dynamic

    ad $ptr, $size, <align>

Allocates memory with dynamic duration (ie, on heap) and returns pointer
to the allocated memory area. The memory is reclaimed automatically
unless the ownership of the pointer is passed to the caller.

The parameters' meaning is the same as for AA.

PTR - PoinTeR

    ptr $ptr, $area, <offset>

Create a pointer to an inside of a memory area allocated by AA or AD.
This makes it easier to address subobjects. However, the since the
offset is only 16 bits wide, the instruction is only able to address the
first 16 KiB of memory in the $area.

ADDIU and ADD instructions can be used to increase the address (the VM
will automatically enforce bounds check to ensure that memory outside
the original area cannot be accessed this way). To facilitate this
analysis put the greedy prefix on PTR instruction and use the arithmetic
instruction immediately afterwards:

    g.ptr $ptr, $area, 0xffff
    addiu $ptr, $ptr, 0x42

Otherwise the pointer arithmetic will be rejected.