Assemble and disassemble AA and AD instructions
The semantics described in commit
241b755da0ee740d9a08c6423d8094bc4eb463ab need to be amended. The
alignment of the memory area is now described by the unit specified in
the pseudoinstructions name:

 - abx: allocate bytes
 - ahx: allocate half-words
 - awx: allocate words
 - adx: allocate double-words
 - aqx: allocate quad-words

The same rule is used for AA and AD instructions. In source code, they
will appear like this:

    aha $ptr, $size, 0
    aa  16, $ptr, $size, 0

The last parameter MUST be zero; otherwise, the source is rejected and
compilation is aborted.

The design for PTR instruction should be revisited.