Remove boxed values from registers
The registers are now able to store fundamental types only. Boxed values
do not exist anymore. There can be built in memory, and accessed through
pointers (which are a fundamental type) placed in registers.

This change requires a BIG rethink of how the VM will treat values and
objects manipulated by the programs it runs.

Previously pretty much everything was boxed (and it was an exception to
be unboxed) and the instructions had to handle boxed and unboxed
varieties of all types. This was unwieldy and forced a number of hard to
maintain constructions on the code. It also put heavy emphasis on small,
frequent heap allocations.

Now, the VM is almost unable to create a boxed value. Structs, buffers,
strings were removed from the set of built-in types and must now be
constructed manually (or by a compiler) using fundamental types and
memory access instructions (AA, SM, LM).

Boxed values will be available through pointers to foreign objects, but
this is still a TODO. "Boxed" may be a little misused here, because the
objects do not necessarily have to be wrapped in some special container
that the VM will understand. They may be opaque pointers to operating
system types eg, sockaddr*, and the only thing that's really needed is a
way to store them in VM's registers and memory.