Show more information in the abort report
The VM will now provide the reason why execution was aborted, the IP
where the failure occured, and the encoded instruction that caused the
failure.

Tests can be written against aborts by putting a foo.abort file beside
the foo.asm test program. The format of these .abort files is:

    <aborted-ip>
    <encoded-instruction>
    <message>