Fetch data from memory as "raw" data instead of uint
The VM has memory store and load instructions based in raw memory, not
on types. This facilitates moves of foreign-allocated values, but makes
it more difficult to determine what kind of value (ie, type of object)
was just loaded into a register.

I tried loading the memory as raw data and then silently converting it
to the "desired" data on on first access, but this proved to be too
brittle. The way to go, I think, will be to provide an instruction to
cast the raw memory into the desired type. Something like the following
would then be used to load a 32-bit uint into a register:

    g.lw $val.l, $addr.l, 0     ; load a word into a register
    cast_to $val.l, ui          ; convert the raw data into an ui

The VM will always treat the cast as valid. The assembler, and a
hypothetical compiler, should not be similarly lax. They will usually
have a sufficient amount of information to properly track the stores and
loads and ensure consistent typing of values.