Add CAST instruction
CAST is used to convert raw memory loaded into a register to a value of
a fundamental type (eg, an integer, a PID, or an atom). Since the VM has
a range of fundamental types and it does not track the "types" of
objects stored in various areas of memory (the VM only tracks sizes of
and offsets inside allocated memory) it is not immediately obvious how
to treat the raw memory that was just loaded.

The solution to this is to explicitly tell the VM what to do with such
memory. For example, to load a float into a register you would do
something like this:

    g.lw $flt.l, $addr.l, 0
    cast $flt.l, float

where "dst" and "addr" are registers containing a float and the address
of that object in memory, respectively.

Of course, after loading the float in a register and the float in memory
are different objects and to update the value in memory you would have
to execute a SM instruction:

    sw $flt.l, $addr.l, 0

This time, however, no cast is needed since the VM has all necessary
information.