Fix SM for non-byte units
Making the stack grow downwards created a bug in the implementation of
the SM instruction. The addresses of the allocated memory were reversed.

Ascending "physical" addresses as seen by C++ corresponded with
descending addresses as seen by user code. This did not play nice with
memcpy(3), but the solution was simple - just remove the
ascending/descending disparity. Now that the addresses allocated by AA
consistently descend, but addresses inside the allocated chunks ascend
eveyrhting's right again.