Implement ARODP instruction
It is to be used to load data from the .rodata section. The VM supplies
the address of the .rodata implicitly, and the user code gives the
offset into .rodata where the desired data lies.

To see ARODP in action, read the source in string.asm test.

Pointers returned by ARODP are "foreign". They point to memory that
wasn't allocated by user code, which means that currently the VM can't
fully guarantee that access to them is safe (as it can with memory
allocated using AA instructions). The guarantee will be there once the
VM becomes able to use struct layout descriptions (eg, DWARF).