User code can echo stdin on stdout
The I/O implementation is crude, basic, and brutal, but for now
entirely sufficient. Basically everything about it is bad, though, and
will have to be redone - except the model.

Submitting I/O requests, letting them be processed asynchronously, and
receiving their results as three separate things is the way to go.