Use LUI+LLI to load long immediates
This change lowers the cost of loading addresses and long integers
significantly, and makes the process more deterministic (either 1 or 2
instructions are used). The previous algorithm which required
multiplication, addition, and several instructions worked, but was a wee
bit ridiculous.

The width of LUI's immediates is reduced from 32 bits to 36. This
ensures symmetry between LUI and LLI, which aids in writing correct
code. Loading 64 bit values is simple - just load the upper and lower
half of the doubleword. Loading longer values is similarly simple:

 - load upper half
 - load lower half

If necessary, split each step into further halves until you get to
32-bit wide values. Irregular integers (eg, 96-bit wide ones) can be
loaded using LUI, LLI, and bit shifting since LLI does not reset the
upper bits of the register it modifies.