Fix an off-by-one error in accesses to the actors
The counting-from-the-top confused me a wee bit too much, I guess.
Accessing memory near the top of the allocated area (precisely - the
first allocated byte) was referencing one byte *past* the page do to
incorrect math.

    (address + size) - 0

does not give 255 if the allocated size is 256 - it gives 256, and there
is no 256th byte in an array sized N. Same for any N and N-1, but 256 is
the size of a page right now.

Anyway, the error is obvious once you notice it.

Thanks again, Valgrind.