Update asm and dis to use .symtab and .strtab
This is a follow-up to commit 6092498fa28ef63fea57da21f88a158c4a84d1ae

Instead of the old code relying on record format specific to the VM,
symbol addresses (both for objects and functions) are now recorded in
the standard .symtab and .strtab sections. This will make ELF files
produced by Viua readable by the standard tools, and ensure that the way
symbols are recorded is actually well defined - we just use the same
standard as (almost) everyone else.

However, this required a rewrite of HUGE swathes of code. Hence the BIG,
UGLY commit.