Update vm to use .symtab and .strtab
This is a follow-up to commit 8186ef252c3ec95206328af2310b3b7ab40bb94b

As with asm and dis, the vm also had to be updated to correctly load
addresses of functions and objects from .rodata section.

What is a bit messed up is that right now there are THREE ways of
loading addresses:

 - functions: are fetched by index in .symtab and translated into
   addresses at run time
 - atoms, floats (and other built-in literals): are fetched by address
   of the DATA in .rodata
 - ARODP: fetches by addresses of objects in .rodata, but produces
   pointers to SIZE prefixing the data

The ARODP will have to be fixed, because when Viua finally allows
storing structs and arbitrary data in .rodata the assumption that
every object is a string prefixed by its size will no longer hold.