Provide both quick and direct access to .strtab
Long strings are usually inserted into the string table "on their own",
and not as part of another string. Access to them is quick through the
cached map.

Shorter strings can be embedded as part of longer strings *if* the
shorter one matches the suffix of the longer one. For example:

 - foobar
 - foo
 - bar

Given the above list of strings the .strtab will look like this:

    \0foobar\0foo\0\0
         ^^^^^

There is no separate entry for "bar" because it was embedded in the
underlined part of "foobar". The same trick could not be employed for
"foo" because a string in .strtab MUST be nul-terminated, and "foo" is
followed by a "b" instead of "\0".

This commit also fixes the deduplication algorithm so that the
assumption outlined in the previous paragraph actually holds. See the
content_matches and nul_matches part.