Add a simple linker
The linker is able to resolve relocations inside a single file, turning
it from a relocatable into executable ELF. That is the ONLY thing it is
so far capable of, but it ensures that the basic compilation process
works at is should work:

 - assembler takes in .asm (source code) and spits out .o (ET_REL) files
 - linker takes in .o files and spits out a.out (ET_EXEC) files
 - interpreter takes in .out files

The test suite passes after the changes.

Gluing together multiple files will come later.