Allow marking functions as externally defined
Use [[extern]] attribute for this purpose:

    .function: [[extern]] example_fn

The assembler will record such a symbol in .symtab, but will not provide
address for it (ie, the st_value will be zero). The linker and the
disassembler can both use this as a signal that this is an externally
defined signal - this should work for both functions and objects.