Test suite handles multi-file programs
The main program is always stored in the file with .asm extension (the
file may be empty); the auxiliary parts of the final program are stored
in files with .s extension.

Every file is assembler, and then they are all linked together. The
first file on the linker's command line is the main one, and the order
of auxiliary files is not defined. This does not matter, though, because
the linker will handle any order of files - provided all the necessary
symbols are defined in the end.