Log commands used to compile each test
They are useful when debugging assembler and linker errors, as I can see
what exactly was invoked. Sure, the output includes the most recent
command that led to the error report, but having the whole history
available gives a fuller picture.

And it can also be used to replay the test:

    ]$ bash ./tests/asm/foo.log

Et voila! It should be able to reproduce the error.