Fix by-name relocations
They are recorded and identified using their offset into the .text
section; but while the list of relocations to apply contained the
correct value, the map used to determine the type of relocation (either
by-index if the symbol was already defined when the relocation was
encountered, or by-name if it was not) contained wrong value.

Simple bug, simple fix.