Wrap funky names in quotes to be safe
If you want to show off and add some class to the names of your
functions - go ahead; the assembler, linker, etc are not here to judge.
They will insist that such names are wrapped in quotes, exactly as
strings are. For example

    .function: std::memcpy
    call void, std::memcpy

are not legal lines. To make them acceptable use quotes:

    .function: "std::memcpy"
    call void, "std::memcpy"

This is only a lexical convention, in the ELF symbol and string tables
the names are recorded without the quotes, and the VM will present the
names stripped in backtraces.