Provide a way to add "dependencies" to test cases
The "dependencies" are just names of modules to link to the main module
of the test case. Instead of reassembling same piece of code over and
over again (eg, implementation of the memcpy(3) function) the test suite
will assemble it once, and then link it to all cases which state it as
their dependency.