Do not allow greedy IF
And, by extension, do not allow greedy jump pseudoinstructions.

Same as CALL instructions cannot be greedy. Flow control instructions
will be treated as points where the scheduler can always break execution
and transfer to another process (of course, it could do it anyway, but
making this assumption explicit is better).