Lex and parse new-style sources
Instead of being semi-structured, the new style is just a flat sequence
of directives and instructions. There are no explicit value and function
definitions:

 - objects are pushed to .rodata
 - instructions are pushed to .text

Labels are used to refer to values inside these two sections; but again,
the labels are not defined by special directives - simple .label just
records current offset into the currently active section.

Structure will be provided by a higher-level language.

FIXME NOTES FOR ERROR MESSAGE RENDERING

The capabilities of the error printer are limited, and not even all the
features that should be there are implemented. Quality of life will be
improved, but only after the basics are working.