Only use minimal warnings on Sourcehut builds
Alpine and NixOS are interesting (the former because it allows me to
test against musl; the latter because of its exotic ways), but they ship
versions of GCC and Clang which are behind the bleeding edge.

A full set of warning flags is suitable for local development on Arch.
There I always have the newest Clang and GCC available and can fiddle
with warnings quickly. Let CI use the minimal reasonable settings and
call it a day.