CAST instruction only accepts undefined type as input
For casting between arithmetic types use arithmetic instructions, as
they automatically cast rhs-value to match the type of lhs-value, and
produce value of the lhs-type. For example, to cast a float to a signed
integer use:

    ; Use zero as lhs to just "copy" rhs.
    ; Then, overwrite the zero with rhs but converted to lhs-type.
    li $2.l, 0
    add $2.l, $2.l, $some_float

See the cast_*.asm files in tests for more examples.

Casting between PID, atom, etc does not really make sense so is not
supported directly. You can always use memory to launder your values,
though:

 - allocate enough memory to hold a value of type T1
 - store a value of type T1 in the allocated memory
 - load a value from memory, it is not of undefined type
 - convert the value of an undefined type to type T2

It works, but is a bit unwieldy. Besides, the VM tracks pointers and
could automatically convert values on load, preventing any laundering
from happening. The only problem with this is foreign code - it can also
allocate and mutate memory so not everything is under the VM's control.

In some cases the VM must just take the cast as correct when the
programmer says "Trust me, I am an engineer - it is alright".