Hide -Wunused errors on NixOS
This not a real fix since the code just loops until it either wrote or
read everything, but without ANY kind of error checks. It will be easy
to fix it later, though, as it will only require changing
implementations of whole_read() and whole_write() instead of ad-hoc
fixes in various places.