Implement checks of data emitted on stdout
To check a program's stdout create a "foo.stdout" file. It must contain
the expected output of the program. The test suite will pick it up and
ensure that the program produced the expected data. The check will be
EXACT so you have to be careful about whitespace, newlines, etc.