Refactor test-case function
The number of source lines did not decrease much, but this is because
the code is now better documented. Logic lines were cut, and comment
lines were added.

This should make it easier to understand what is happening during the
next refactoring, or when I need to add some features to the test suite.