Automatically set version and date in man pages during installation
This is much better than relying on manual updates. Now, the
installation command

    ]$ make install

will sed(1) the manual files to put proper date and version in them.

The assumption is that you are always installing from HEAD, as this is
what the make_version.sh script uses. We need to follow the same logic
in the Makefile to keep date and version consistent.